The FINAL TOUCH

+++ src/controllers/MyWallpapersController.java
@@ -13,14 +13,18 @@ import java.util.List;
 import java.util.ResourceBundle;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import javafx.collections.ListChangeListener;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
+import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.TilePane;
 import javafx.scene.layout.VBox;
+import utils.ImagesUtils;
+import utils.Toast;
 
 public class MyWallpapersController implements Initializable {
 
@@ -36,9 +40,21 @@ public class MyWallpapersController implements Initializable {
     private JFXButton saveImageToHDButton;
     @FXML
     private JFXButton removeImageFromHDButton;
+    @FXML
+    private ImageView noWallpaperFoundIMageView;
 
     @Override
     public void initialize(URL url, ResourceBundle rb) {
+
+        wallpapersTilePane.getChildren().addListener((ListChangeListener.Change<? extends Node> c) -> {
+            if (c.next()) {
+                if (c.wasRemoved()) {
+                    if (wallpapersTilePane.getChildren().isEmpty()) {
+                        noWallpaperFoundIMageView.setVisible(true);
+                    }
+                }
+            }
+        });
 
         try (Stream<Path> walk = Files.walk(Paths.get(new File("src/../wallpapers").toURI()))) {
 
@@ -56,9 +72,9 @@ public class MyWallpapersController implements Initializable {
                 w.setImageURL(imageLocation);
 
                 w.setImageURLThumbnail(imageLocation);
-                
+
                 w.setInMyWallpaper(true);
-                
+
                 w.setTilePaneParent(wallpapersTilePane);
 
                 w.init(choosenImageVBox, spinner, choosenImageView, saveImageToHDButton);
@@ -67,6 +83,10 @@ public class MyWallpapersController implements Initializable {
             }
         } catch (IOException e) {
             e.printStackTrace();
+        }
+
+        if (wallpapersTilePane.getChildren().isEmpty()) {
+            noWallpaperFoundIMageView.setVisible(true);
         }
     }
 
@@ -77,17 +97,28 @@ public class MyWallpapersController implements Initializable {
 
     @FXML
     private void removeImageFromMyWallappers(ActionEvent event) throws IOException {
-        choosenImageView.setImage(null);
 
-        AppData.choosenWallpaper.removeWallpaper(wallpapersTilePane);
+        try {
+            if (Toast.showConfirmation()) {
 
-        Files.deleteIfExists(Paths.get(AppData.choosenWallpaper.getImageURL().substring(6)));
+                Files.deleteIfExists(Paths.get(AppData.choosenWallpaper.getImageURL().substring(6)));
 
-        choosenImageVBox.setVisible(false);
+                choosenImageView.setImage(null);
+
+                AppData.choosenWallpaper.removeWallpaper(wallpapersTilePane);
+
+                choosenImageVBox.setVisible(false);
+
+                Toast.makeText(lunch.Lunch.appStage, "Wallpaper Removed Succssfully", 1500);
+            }
+        } catch (IOException e) {
+            Toast.makeText(lunch.Lunch.appStage, "Couldn't Remove Wallpaper, Try Agian Later", 1500);
+        }
+
     }
 
     @FXML
     private void saveImageToHD(ActionEvent event) {
-        
+        ImagesUtils.saveImageAsToHD();
     }
 }
